Added javascript syntax extensions for highlighting missing semicolons.
It's not perfect, and it errs on the side of missing some scenarios rather than
highlighting places where semicolons aren't really needed.

For instance, the last statement in a block should have a semicolon, but it
won't be highlighted.  This is because the exception for the last line in a
object literal shouldn't have a semicolon - but there's no good way for me to
determine which context we are in (there's no region for that (yet)).

I will experiment with creating syntax regions for functions and literals.  If
that works out, then I should be able to add literals to the ALLBUT list.