Prefixed TmuxRunCommand string with C-c to clear any pre-existing junk on the target pane.
For example, sometimes I accidentally press some keys intended for vim while the
focus is actually on another pane, and I'll leave something like "jjkkk" on the
command line.  Then when I send-keys a command over to tmux, it looks something
like "jjkkkclear; tmux clear-history; ..." - then bash complains about not
knowing what the jjkkkclear command is.  And I have to see at it.  This is bad.